Move mod-specific lint attributes to Mods.Common.
