+++ OpenRA.Mods.Common/Lint/CheckActorReferences.cs
@@ -13,6 +13,7 @@ using System;
 using System.Linq;
 using System.Reflection;
 using OpenRA.GameRules;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.Common.Lint
